requirements.txt
requests>=2.31.0
pinecone-client>=2.2.4
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=10.1.0
PyMuPDF>=1.23.7
openai>=1.3.7
numpy>=1.26.2
tqdm>=4.66.1
streamlit>=1.29.0
python-dotenv>=1.0.0
boto3>=1.34.0 